新增推送

+++ app/src/main/java/com/whmnrc/flymall/beans/HomeSaleGoodsBean.java
@@ -135,9 +135,9 @@ public class HomeSaleGoodsBean {
             private boolean HasSKU;
             private int Id;
             private String ImagePath;
-            private int MarketPrice;
+            private double MarketPrice;
             private String MeasureUnit;
-            private int MinSalePrice;
+            private double MinSalePrice;
             private String ProductCode;
             private String ProductName;
             private int Quantity;
@@ -262,11 +262,11 @@ public class HomeSaleGoodsBean {
                 this.ImagePath = ImagePath;
             }
 
-            public int getMarketPrice() {
+            public double getMarketPrice() {
                 return MarketPrice;
             }
 
-            public void setMarketPrice(int MarketPrice) {
+            public void setMarketPrice(double MarketPrice) {
                 this.MarketPrice = MarketPrice;
             }
 
@@ -278,11 +278,11 @@ public class HomeSaleGoodsBean {
                 this.MeasureUnit = MeasureUnit;
             }
 
-            public int getMinSalePrice() {
+            public double getMinSalePrice() {
                 return MinSalePrice;
             }
 
-            public void setMinSalePrice(int MinSalePrice) {
+            public void setMinSalePrice(double MinSalePrice) {
                 this.MinSalePrice = MinSalePrice;
             }
